+++ CHANGES.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2025-12-2x
+
+### Breaking
+
+- `resources.default.path_maker` is now `pm_file` instead of `pm_class`. This only affects resources relative to a test `some_package.TestSomeThing.test_an_aspect` where the base path is now `some_package/TestSomeThing__test_an_aspect` instead of `some_package__TestSomeThing/test_an_aspect`. This gives fewer directories next to the test file and less clutter.
+
 ## [0.9.0] - 2025-10-21
 
 ### Added